// client/palette.ts
export interface Palette {
    black: string;
    darkBlue: string;
    darkPurple: string;
    darkGreen: string;
    brown: string;
    darkGrey: string;
    lightGrey: string;
    white: string;
    red: string;
    orange: string;
    yellow: string;
    green: string;
    blue: string;
    lavender: string;
    pink: string;
    lightPeach: string;
}

export const PICO_8_PALETTE: Palette = {
    black: "#000000",
    darkBlue: "#1D2B53",
    darkPurple: "#7E2553",
    darkGreen: "#008751",
    brown: "#AB5236",
    darkGrey: "#5F574F",
    lightGrey: "#C2C3C7",
    white: "#FFF1E8",
    red: "#FF004D",
    orange: "#FFA300",
    yellow: "#FFEC27",
    green: "#00E436",
    blue: "#29ADFF",
    lavender: "#83769C",
    pink: "#FF77A8",
    lightPeach: "#FFCCAA",
};